# platform/pkg/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.22.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Live struct {
	LiveID             int32
	UserID             pgtype.UUID
	Title              string
	Description        pgtype.Text
	StartTime          pgtype.Timestamp
	EndTime            pgtype.Timestamp
	ScheduledStartTime pgtype.Timestamp
	ScheduledEndTime   pgtype.Timestamp
	LiveAppName        pgtype.Text
	StreamName         pgtype.Text
	LiveSecret         []byte
	StreamBroadcastUrl []byte
	CreatedAt          pgtype.Timestamptz
	UpdatedAt          pgtype.Timestamptz
}

type LiveLog struct {
	LiveLogID      int32
	LiveID         pgtype.Int4
	Action         pgtype.Text
	ClientID       pgtype.Text
	StreamID       pgtype.Text
	ServerID       pgtype.Text
	ServiceID      pgtype.Text
	Ip             pgtype.Text
	Vhost          pgtype.Text
	App            pgtype.Text
	Tcurl          pgtype.Text
	StreamUrlParam pgtype.Text
	CreatedAt      pgtype.Timestamptz
	UpdatedAt      pgtype.Timestamptz
}

type LiveStat struct {
	LiveStatID int32
	LiveID     pgtype.Int4
	Status     pgtype.Text
	CreatedAt  pgtype.Timestamptz
	UpdatedAt  pgtype.Timestamptz
}

type User struct {
	UserID        pgtype.UUID
	FirebaseUid   pgtype.Text
	Email         string
	EmailVerified pgtype.Bool
	FirstName     pgtype.Text
	LastName      pgtype.Text
	DisplayName   pgtype.Text
	PhotoUrl      pgtype.Text
	PhoneNumber   pgtype.Text
	CreatedAt     pgtype.Timestamptz
	UpdatedAt     pgtype.Timestamptz
}